Skip null inputstream if requested resource is not part of jar
git-svn-id: svn+ssh://svn.msr-inria.inria.fr/var/lib/svn/repository/tla/trunk/tla2-inria@22277 76a6fc44-f60b-0410-a9a8-e67b0e8fc65c

+++ tlatools/src/util/SimpleFilenameToStream.java
@@ -156,23 +156,25 @@ private static String installationBasePath = getInstallationBasePath();
 					InputStream is = cl
 							.getResourceAsStream(STANDARD_MODULES
 									+ name);
-        		
-        		try {
-					sourceFile = new File(TMPDIR + File.separator + name);
-					sourceFile.deleteOnExit();
 					
-					FileOutputStream fos = new FileOutputStream(sourceFile);
-					
-					byte buf[] = new byte[1024];
-					int len;
-					while ((len = is.read(buf)) > 0) {
-						fos.write(buf, 0, len);
+				if(is != null) {
+					try {
+						sourceFile = new File(TMPDIR + File.separator + name);
+						sourceFile.deleteOnExit();
+						
+						FileOutputStream fos = new FileOutputStream(sourceFile);
+						
+						byte buf[] = new byte[1024];
+						int len;
+						while ((len = is.read(buf)) > 0) {
+							fos.write(buf, 0, len);
+						}
+						fos.close();
+						is.close();
+					} catch (IOException e) {
+						// must not happen
+						e.printStackTrace();
 					}
-					fos.close();
-					is.close();
-				} catch (IOException e) {
-					// must not happen
-					e.printStackTrace();
 				}
         	} else {
         		sourceFile = new File( prefix + name );
